requirements.txt
# JalDrishti — Maritime Security Vision System
# ─────────────────────────────────────────────
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.21.0
scipy>=1.11.0
matplotlib>=3.7.0

# Web API
flask>=3.0.0
flask-cors>=4.0.0

# ─── Optional (install for full capability) ──────────────────────────────────
# Object Detection
ultralytics>=8.0.0          # pip install ultralytics

# Depth estimation (MiDaS uses timm)
timm>=0.9.0                 # pip install timm

# Video processing
opencv-python>=4.8.0        # pip install opencv-python

# Training Upgrades
lightning>=2.1.0
torchmetrics>=1.2.0
segmentation-models-pytorch>=0.3.3
tensorboard>=2.14.0
